requirements.txt
# AI Grading System - 国产大模型简答题评分系统
# 使用国产大模型替代国外模型

# Web Framework
Flask==3.0.2
Flask-CORS==4.0.0

# 国产大模型 SDK
dashscope==1.14.0          # 通义千问 (阿里云)
zhipuai==2.0.1              # 智谱 GLM
qianfan==0.3.0              # 百度文心一言

# 数据处理
pydantic==2.6.3
pydantic-settings==2.1.0
numpy==1.26.4
pandas==2.2.1

# 异步支持
aiohttp==3.9.3
asyncio==3.4.3

# 工具库
python-dotenv==1.0.1
PyYAML==6.0.1
loguru==0.7.2

# 数据库
SQLAlchemy==2.0.28

# 测试
pytest==8.1.1
pytest-asyncio==0.23.6
